Split out Validation and Compression
